add regexp filename parser, refactor filename parsing code for testability
