+++ pyplis/setupclasses.py
@@ -37,7 +37,8 @@ from .forms import LineCollection, RectCollection
 from .helpers import isnum, to_datetime
 from .exceptions import MetaAccessError, DeprecationError
 from .inout import get_source_info, save_default_source
-from .utils import Filter, CameraBaseInfo
+from .utils import Filter
+from .CameraBaseInfo import CameraBaseInfo
 from .geometry import MeasGeometry
